+++ sdk/textanalytics/azure-ai-textanalytics/azure/ai/textanalytics/_generated/v3_1_preview_3/models/_models.py
@@ -2028,7 +2028,7 @@ class TasksStateTasks(msrest.serialization.Model):
      list[~azure.ai.textanalytics.v3_1_preview_3.models.TasksStateTasksEntityRecognitionPiiTasksItem]
     :param key_phrase_extraction_tasks:
     :type key_phrase_extraction_tasks:
-     list[~azure.ai.textanalytics.v3_1_preview_3.models.TasksStateTasksKeyPhraseExtractionTasksItem]
+     list[~azure.ai.textanalytics.v3_1_preview_3.models.TasksStateTasksKeyPhraseExtractionActionsItem]
     """
 
     _validation = {
@@ -2046,7 +2046,7 @@ class TasksStateTasks(msrest.serialization.Model):
         'total': {'key': 'total', 'type': 'int'},
         'entity_recognition_tasks': {'key': 'entityRecognitionTasks', 'type': '[TasksStateTasksEntityRecognitionTasksItem]'},
         'entity_recognition_pii_tasks': {'key': 'entityRecognitionPiiTasks', 'type': '[TasksStateTasksEntityRecognitionPiiTasksItem]'},
-        'key_phrase_extraction_tasks': {'key': 'keyPhraseExtractionTasks', 'type': '[TasksStateTasksKeyPhraseExtractionTasksItem]'},
+        'key_phrase_extraction_tasks': {'key': 'keyPhraseExtractionTasks', 'type': '[TasksStateTasksKeyPhraseExtractionActionsItem]'},
     }
 
     def __init__(
@@ -2211,8 +2211,8 @@ class TasksStateTasksEntityRecognitionTasksItem(TaskState, Components15Gvwi3Sche
         self.status = kwargs['status']
 
 
-class TasksStateTasksKeyPhraseExtractionTasksItem(TaskState, Components1D9IzucSchemasTasksstatePropertiesTasksPropertiesKeyphraseextractiontasksItemsAllof1):
-    """TasksStateTasksKeyPhraseExtractionTasksItem.
+class TasksStateTasksKeyPhraseExtractionActionsItem(TaskState, Components1D9IzucSchemasTasksstatePropertiesTasksPropertiesKeyphraseextractiontasksItemsAllof1):
+    """TasksStateTasksKeyPhraseExtractionActionsItem.
 
     All required parameters must be populated in order to send to Azure.
 
@@ -2244,7 +2244,7 @@ class TasksStateTasksKeyPhraseExtractionTasksItem(TaskState, Components1D9IzucSc
         self,
         **kwargs
     ):
-        super(TasksStateTasksKeyPhraseExtractionTasksItem, self).__init__(**kwargs)
+        super(TasksStateTasksKeyPhraseExtractionActionsItem, self).__init__(**kwargs)
         self.results = kwargs['results']
         self.last_update_date_time = kwargs['last_update_date_time']
         self.name = kwargs.get('name', None)